fix: :recycle: break apart the encryption from the storage
